put train, evaluate and infer functions in core
