+++ dygraph/val.py
@@ -32,6 +32,7 @@ import utils.logging as logging
 from utils import get_environ_info
 from utils import ConfusionMatrix
 from utils import Timer, calculate_eta
+from core import evaluate
 
 
 def parse_args():
@@ -71,65 +72,6 @@ def parse_args():
         default=None)
 
     return parser.parse_args()
-
-
-def evaluate(model,
-             eval_dataset=None,
-             model_dir=None,
-             num_classes=None,
-             ignore_index=255,
-             epoch_id=None):
-    ckpt_path = os.path.join(model_dir, 'model')
-    para_state_dict, opti_state_dict = fluid.load_dygraph(ckpt_path)
-    model.set_dict(para_state_dict)
-    model.eval()
-
-    total_steps = len(eval_dataset)
-    conf_mat = ConfusionMatrix(num_classes, streaming=True)
-
-    logging.info(
-        "Start to evaluating(total_samples={}, total_steps={})...".format(
-            len(eval_dataset), total_steps))
-    timer = Timer()
-    timer.start()
-    for step, (im, im_info, label) in enumerate(eval_dataset):
-        im = to_variable(im)
-        pred, _ = model(im)
-        pred = pred.numpy().astype('float32')
-        pred = np.squeeze(pred)
-        for info in im_info[::-1]:
-            if info[0] == 'resize':
-                h, w = info[1][0], info[1][1]
-                pred = cv2.resize(pred, (w, h), cv2.INTER_NEAREST)
-            elif info[0] == 'padding':
-                h, w = info[1][0], info[1][1]
-                pred = pred[0:h, 0:w]
-            else:
-                raise Exception("Unexpected info '{}' in im_info".format(
-                    info[0]))
-        pred = pred[np.newaxis, :, :, np.newaxis]
-        pred = pred.astype('int64')
-        mask = label != ignore_index
-
-        conf_mat.calculate(pred=pred, label=label, ignore=mask)
-        _, iou = conf_mat.mean_iou()
-
-        time_step = timer.elapsed_time()
-        remain_step = total_steps - step - 1
-        logging.info(
-            "[EVAL] Epoch={}, Step={}/{}, iou={:4f}, sec/step={:.4f} | ETA {}".
-            format(epoch_id, step + 1, total_steps, iou, time_step,
-                   calculate_eta(remain_step, time_step)))
-        timer.restart()
-
-    category_iou, miou = conf_mat.mean_iou()
-    category_acc, macc = conf_mat.accuracy()
-    logging.info("[EVAL] #image={} acc={:.4f} IoU={:.4f}".format(
-        len(eval_dataset), macc, miou))
-    logging.info("[EVAL] Category IoU: " + str(category_iou))
-    logging.info("[EVAL] Category Acc: " + str(category_acc))
-    logging.info("[EVAL] Kappa:{:.4f} ".format(conf_mat.kappa()))
-    return miou, macc
 
 
 def main(args):
